backend/recipes/requirements.txt
Django==5.2.6
google-genai==1.39.1
Pillow==11.3.0
python-dotenv==1.1.1
django-cors-headers==4.6.0